+++ src/Common/Commands/StackPanelChangeVisibilityCommand.cs
@@ -1,9 +1,8 @@
 ﻿using System;
 using System.Windows.Controls;
 using System.Windows.Input;
-using Unchase.OpenAPI.ConnectedService.Common;
 
-namespace Unchase.OpenAPI.ConnectedService.Commands
+namespace Unchase.OpenAPI.ConnectedService.Common.Commands
 {
     class StackPanelChangeVisibilityCommand : ICommand
     {